requirements.txt
pandas
numpy
scikit-learn
indic-nlp-library
stopwordsiso
faiss-cpu
transformers
torch
sentence-transformers
tqdm
tiktoken
flask
flask-cors
protobuf